use react-router to manage tabs
