+++ components/tabs/TabUrlMap.jsx
@@ -1,0 +1,9 @@
+const tabNameMap = {
+  WorkRecord : "/WorkRecord",
+  Material : "/Material",
+  Voters : "/Voters",
+  ApiDoc : "/ApiDoc",
+  Login : "/Login"
+}
+
+export default tabNameMap
