+++ components/tabs/index.js
@@ -1,0 +1,7 @@
+export { default as Login } from './Login.jsx'
+export { default as ApiDoc } from './ApiDoc.jsx'
+export { default as Voters } from './Voters.jsx'
+export { default as WorkRecord } from './WorkRecord.jsx'
+export { default as Material } from './SideBar.jsx'
+
+export { default as TabUrlMap } from './TabUrlMap.jsx'
